Cleanup iterating vectors and some inheritance (#4)
* Cleanup iterating vectors

* Use unique ptr for powerup

* Use smart pointers for barriers

* Change coordinate type to float

* Make width and height a float

* Add pointer to auto types

* Remove some implicit casts to int

* Fix some general code smells

+++ src/game/bullet.cpp
@@ -13,8 +13,8 @@ Bullet::Bullet(World* world,
   this->worldPointer = world;
   this->x = x;
   this->y = y;
-  this->vector_x = -speed * cos(angle);
-  this->vector_y = -speed * sin(angle);
+  this->vector_x = -speed * cosf(angle);
+  this->vector_y = -speed * sinf(angle);
   this->shotSound = sound;
   play_sample(shotSound, 255, 127, random(800, 1200), 0);
   this->pendingErase = false;
